Add kits for OMI version v1.2.0-35
Note that HP and Solaris kits are compressed, and must be uncompressed
prior to using them natively.

The 'compress' program was chosen for these platforms as that's what is
natively shipped by the manufacturer without any special package installs.

Further note that Solaris kits appear to be built DEBUG (inadvertantly).
I suspect that HP is as well (based on the size), although I haven't
verified that. We will verify all platform builds in an upcoming release.